+++ src/adapter/src/coord/sequencer/cluster.rs
@@ -31,6 +31,7 @@ use mz_sql::plan::{
 };
 use mz_sql::session::metadata::SessionMetadata;
 use mz_sql::session::vars::{SystemVars, Var, MAX_REPLICAS_PER_CLUSTER};
+use mz_sql_parser::ast::ClusterScheduleOptionValue;
 
 use crate::catalog::Op;
 use crate::coord::Coordinator;
@@ -70,6 +71,7 @@ impl Coordinator {
                     replication_factor: plan.replication_factor,
                     disk: plan.disk,
                     optimizer_feature_overrides: plan.optimizer_feature_overrides.clone(),
+                    schedule: plan.schedule.clone(),
                 })
             }
             CreateClusterVariant::Unmanaged(_) => ClusterVariant::Unmanaged,
@@ -108,6 +110,7 @@ impl Coordinator {
             size,
             disk,
             optimizer_feature_overrides: _,
+            schedule,
         }: CreateClusterManagedPlan,
         cluster_id: ClusterId,
         mut ops: Vec<catalog::Op>,
@@ -134,6 +137,14 @@ impl Coordinator {
             "cluster replica",
             MAX_REPLICAS_PER_CLUSTER.name(),
         )?;
+
+        if !matches!(schedule, ClusterScheduleOptionValue::Manual) {
+            // todo: remove this check once ClusterScheduleOptionValue::Refresh is
+            // actually implemented.
+            return Err(AdapterError::Unsupported(
+                "cluster schedules other than MANUAL",
+            ));
+        }
 
         for replica_name in (0..replication_factor).map(managed_cluster_replica_name) {
             let id = self.catalog_mut().allocate_user_replica_id().await?;
@@ -579,6 +590,7 @@ impl Coordinator {
                     replication_factor: 1,
                     disk,
                     optimizer_feature_overrides: Default::default(),
+                    schedule: Default::default(),
                 });
             }
         }
@@ -592,6 +604,7 @@ impl Coordinator {
                 replication_factor,
                 disk,
                 optimizer_feature_overrides: _,
+                schedule,
             }) => {
                 use AlterOptionParameter::*;
                 match &options.size {
@@ -627,6 +640,20 @@ impl Coordinator {
                 match &options.replication_factor {
                     Set(rf) => *replication_factor = *rf,
                     Reset => *replication_factor = 1,
+                    Unchanged => {}
+                }
+                match &options.schedule {
+                    Set(new_schedule) => {
+                        if !matches!(new_schedule, ClusterScheduleOptionValue::Manual) {
+                            // todo: remove this check once ClusterScheduleOptionValue::Refresh is
+                            // actually implemented.
+                            return Err(AdapterError::Unsupported(
+                                "cluster schedules other than MANUAL",
+                            ));
+                        }
+                        *schedule = new_schedule.clone();
+                    }
+                    Reset => *schedule = Default::default(),
                     Unchanged => {}
                 }
                 if !matches!(options.replicas, Unchanged) {
@@ -712,6 +739,7 @@ impl Coordinator {
                 idle_arrangement_merge_effort,
                 disk,
                 optimizer_feature_overrides: _,
+                schedule: _,
             },
             ClusterVariantManaged {
                 size: new_size,
@@ -721,6 +749,7 @@ impl Coordinator {
                 idle_arrangement_merge_effort: new_idle_arrangement_merge_effort,
                 disk: new_disk,
                 optimizer_feature_overrides: _,
+                schedule: _,
             },
         ) = (&config, &new_config);
 
@@ -856,6 +885,7 @@ impl Coordinator {
             idle_arrangement_merge_effort: _,
             disk: new_disk,
             optimizer_feature_overrides: _,
+            schedule: _,
         } = &mut new_config;
 
         // Validate replication factor parameter